Pull abandoned package from Nuts repo
A few years ago, Ben S forked PreMailer from the original repo, and
our builds continued to work after the original author abandoned it,
as it was pulling the package from the forked repo.

Unfortunately, PreMailer itself uses crossorigin/css which hasn't been
forked, and the original repo has been abandoned as well, which broke
our builds.

This updates the composer.json to pull the crossjoin/css package from
the Nuts cloned repo, so that our builds continue to work.

@see https://github.com/nuts/crossjoin_css